+++ src/lib/builders/range-calendar/create.ts
@@ -108,6 +108,8 @@ export function createRangeCalendar<T extends DateValue = DateValue>(
 		defaultPlaceholder: withDefaults.defaultPlaceholder,
 	});
 	const formatter = createFormatter(get(locale));
+	const valueWritable = withDefaults.value ?? writable(withDefaults.defaultValue);
+	const value = overridable(valueWritable, withDefaults.onValueChange);
 
 	const defaultStart = withDefaults.value ? get(withDefaults.value)?.start : undefined;
 	const startValue = writable<DateValue | undefined>(
@@ -116,9 +118,6 @@ export function createRangeCalendar<T extends DateValue = DateValue>(
 
 	const defaultEnd = withDefaults.value ? get(withDefaults.value)?.end : undefined;
 	const endValue = writable<DateValue | undefined>(defaultEnd ?? withDefaults.defaultValue?.end);
-
-	const valueWritable = withDefaults.value ?? writable(withDefaults.defaultValue);
-	const value = overridable(valueWritable, withDefaults.onValueChange);
 
 	const placeholderWritable =
 		withDefaults.placeholder ?? writable(withDefaults.defaultPlaceholder ?? defaultDate);
@@ -448,7 +447,7 @@ export function createRangeCalendar<T extends DateValue = DateValue>(
 				const isOutsideMonth = !isSameMonth(cellValue, monthValue);
 				const isFocusedDate = isSameDay(cellValue, $placeholder);
 				const isOutsideVisibleMonths = $isOutsideVisibleMonths(cellValue);
-				const isSelectedDate = $isSelected(cellValue) && !isOutsideMonth;
+				const isSelectedDate = $isSelected(cellValue);
 				const isSelectionStart = $isSelectionStart(cellValue);
 				const isSelectionEnd = $isSelectionEnd(cellValue);
 				const isHighlighted = $highlightedRange
@@ -994,9 +993,13 @@ export function createRangeCalendar<T extends DateValue = DateValue>(
 	});
 
 	effect([startValue, endValue], ([$startValue, $endValue]) => {
+		const $value = get(value);
+
+		if ($value && $value?.start === $startValue && $value?.end === $endValue) return;
+
 		if ($startValue && $endValue) {
-			valueWritable.update((prev) => {
-				if (prev.start === $startValue && prev.end === $endValue) {
+			value.update((prev) => {
+				if (prev?.start === $startValue && prev?.end === $endValue) {
 					return prev;
 				}
 				if (isBefore($endValue, $startValue)) {
@@ -1011,8 +1014,8 @@ export function createRangeCalendar<T extends DateValue = DateValue>(
 					};
 				}
 			});
-		} else {
-			valueWritable.set({
+		} else if ($value && $value.start && $value.end) {
+			value.set({
 				start: undefined,
 				end: undefined,
 			});
@@ -1034,6 +1037,8 @@ export function createRangeCalendar<T extends DateValue = DateValue>(
 			weekdays,
 			headingValue,
 			value,
+			startValue,
+			endValue,
 		},
 		helpers: {
 			nextPage,